+++ heart-rate-sensor-data-worker/src/main/java/com/sport/heartratesensordataworker/model/UserHeartRate.java
@@ -15,5 +15,5 @@ public class UserHeartRate {
     private String id;
 
     private String userId;
-    private float heartRate;
+    private int heartRate;
 }
